refactor: move annotations to features/annotations
