+++ packages/base/src/features/annotations/components/AnnotationFloater.tsx
