+++ packages/base/src/features/annotations/index.ts
@@ -1,3 +1,4 @@
 export * from './model';
 export * from './components/Annotation';
 export * from './components/AnnotationFloater';
+export * from './AnnotationsPanel';
